fix(shared,indexer): resolve credit review findings
