+++ packages/indexer/src/EventHandlers.ts
@@ -27,3 +27,4 @@ import "./handlers/settlement";
 import "./handlers/settlement-lifecycle";
 import "./handlers/settlement-executor";
 import "./handlers/credit-registry";
+import "./handlers/credit-registry-loans";
