+++ packages/shared/src/hooks/index.ts
@@ -1,5 +1,4 @@
 // Hooks — EXPLICIT EXPORTS for tree-shaking
-
 // ============================================================================
 // QUERY KEYS
 // ============================================================================
