+++ packages/shared/src/index.ts
@@ -206,7 +206,6 @@ export {
   WorkbenchRow,
   WorkCard as WorkCardComponent,
 } from "./components/index";
-
 // ============================================================================
 // CONFIG
 // ============================================================================
