+++ tests/specs/client.offline-sync.spec.ts
@@ -21,7 +21,7 @@ const APP_GARDEN_DETAIL_URL = /\/home\/(?!garden(?:\/|$)|login(?:\/|$)|profile(?
 
 // Skip entire file - these tests require real auth infrastructure
 test.describe("Offline Sync Flows", () => {
-  // SKIP: #338 owner:afo expiry:2026-08-17 — needs real auth infrastructure
+  // SKIP: #338 owner:afo expiry:2026-09-17 — needs real auth infrastructure
   test.skip(
     () => true,
     "Offline sync e2e tests skipped: require real auth infrastructure. " +
@@ -146,7 +146,7 @@ test.describe("Offline Sync Flows", () => {
       await page.waitForLoadState("domcontentloaded");
 
       const gardensAvailable = await hasGardens(page);
-      // SKIP: #338 owner:afo expiry:2026-08-17 — runtime data availability check
+      // SKIP: #338 owner:afo expiry:2026-09-17 — runtime data availability check
       test.skip(!gardensAvailable, "No gardens available for testing");
 
       // Navigate to work submission
@@ -260,7 +260,7 @@ test.describe("Offline Sync Flows", () => {
       await page.waitForLoadState("domcontentloaded");
 
       const gardensAvailable = await hasGardens(page);
-      // SKIP: #338 owner:afo expiry:2026-08-17 — runtime data availability check
+      // SKIP: #338 owner:afo expiry:2026-09-17 — runtime data availability check
       test.skip(!gardensAvailable, "No gardens available for testing");
 
       // Queue work offline
@@ -318,7 +318,7 @@ test.describe("Offline Sync Flows", () => {
       await page.waitForLoadState("domcontentloaded");
 
       const gardensAvailable = await hasGardens(page);
-      // SKIP: #338 owner:afo expiry:2026-08-17 — runtime data availability check
+      // SKIP: #338 owner:afo expiry:2026-09-17 — runtime data availability check
       test.skip(!gardensAvailable, "No gardens available for testing");
 
       // Queue work offline
@@ -411,7 +411,7 @@ test.describe("Offline Sync Flows", () => {
       await page.waitForLoadState("domcontentloaded");
 
       const gardensAvailable = await hasGardens(page);
-      // SKIP: #338 owner:afo expiry:2026-08-17 — runtime data availability check
+      // SKIP: #338 owner:afo expiry:2026-09-17 — runtime data availability check
       test.skip(!gardensAvailable, "No gardens available for testing");
 
       // Queue work offline
@@ -495,7 +495,7 @@ test.describe("Offline Sync Flows", () => {
   test.describe("Conflict Resolution", () => {
     test("detects duplicate submission conflict", async ({ page }) => {
       // This test requires a complex setup - skip for now with explanation
-      // SKIP: #338 owner:afo expiry:2026-08-17 — needs pre-existing duplicate work
+      // SKIP: #338 owner:afo expiry:2026-09-17 — needs pre-existing duplicate work
       test.skip(
         true,
         "Conflict detection requires pre-existing duplicate work - manual testing recommended"
@@ -506,7 +506,7 @@ test.describe("Offline Sync Flows", () => {
 
     test("shows conflict resolution UI", async ({ page }) => {
       // This test requires triggering a conflict state
-      // SKIP: #338 owner:afo expiry:2026-08-17 — needs simulated duplicate
+      // SKIP: #338 owner:afo expiry:2026-09-17 — needs simulated duplicate
       test.skip(true, "Conflict UI requires simulated duplicate - manual testing recommended");
       await page.goto("/home");
       await page.waitForLoadState("domcontentloaded");
@@ -514,7 +514,7 @@ test.describe("Offline Sync Flows", () => {
 
     test("resolves conflict by skipping duplicate", async ({ page }) => {
       // This test requires a conflict to be present
-      // SKIP: #338 owner:afo expiry:2026-08-17 — needs active conflict state
+      // SKIP: #338 owner:afo expiry:2026-09-17 — needs active conflict state
       test.skip(true, "Conflict resolution requires active conflict - manual testing recommended");
       await page.goto("/home");
       await page.waitForLoadState("domcontentloaded");
@@ -549,7 +549,7 @@ test.describe("Offline Sync Flows", () => {
 
     test("warns when storage is nearly full", async ({ page }) => {
       // Storage warning requires filling storage - hard to test in E2E
-      // SKIP: #338 owner:afo expiry:2026-08-17 — needs filled local storage
+      // SKIP: #338 owner:afo expiry:2026-09-17 — needs filled local storage
       test.skip(
         true,
         "Storage warning requires filling local storage - manual testing recommended"
