+++ tests/specs/client.work-submission.spec.ts
@@ -21,7 +21,7 @@ const APP_GARDEN_DETAIL_URL = /\/home\/(?!garden(?:\/|$)|login(?:\/|$)|profile(?
 
 // Skip entire file - these tests require real auth and blockchain infrastructure
 test.describe("Work Submission Flows", () => {
-  // SKIP: #338 owner:afo expiry:2026-08-17 — needs real auth + gardens data
+  // SKIP: #338 owner:afo expiry:2026-09-17 — needs real auth + gardens data
   test.skip(
     () => true,
     "Work submission e2e tests skipped: require real auth and gardens data. " +
@@ -46,7 +46,7 @@ test.describe("Work Submission Flows", () => {
 
       // Check if gardens are available
       const gardensAvailable = await hasGardens(page);
-      // SKIP: #338 owner:afo expiry:2026-08-17 — runtime data availability check
+      // SKIP: #338 owner:afo expiry:2026-09-17 — runtime data availability check
       test.skip(!gardensAvailable, "No gardens available for testing");
 
       // Click first garden card
@@ -81,7 +81,7 @@ test.describe("Work Submission Flows", () => {
       await page.waitForLoadState("domcontentloaded");
 
       const gardensAvailable = await hasGardens(page);
-      // SKIP: #338 owner:afo expiry:2026-08-17 — runtime data availability check
+      // SKIP: #338 owner:afo expiry:2026-09-17 — runtime data availability check
       test.skip(!gardensAvailable, "No gardens available for testing");
 
       // Navigate to work submission
@@ -130,7 +130,7 @@ test.describe("Work Submission Flows", () => {
       await page.waitForLoadState("domcontentloaded");
 
       const gardensAvailable = await hasGardens(page);
-      // SKIP: #338 owner:afo expiry:2026-08-17 — runtime data availability check
+      // SKIP: #338 owner:afo expiry:2026-09-17 — runtime data availability check
       test.skip(!gardensAvailable, "No gardens available for testing");
 
       // Navigate to work submission
@@ -170,7 +170,7 @@ test.describe("Work Submission Flows", () => {
       await page.waitForLoadState("domcontentloaded");
 
       const gardensAvailable = await hasGardens(page);
-      // SKIP: #338 owner:afo expiry:2026-08-17 — runtime data availability check
+      // SKIP: #338 owner:afo expiry:2026-09-17 — runtime data availability check
       test.skip(!gardensAvailable, "No gardens available for testing");
 
       // Navigate to work submission
@@ -225,7 +225,7 @@ test.describe("Work Submission Flows", () => {
       await page.waitForLoadState("domcontentloaded");
 
       const gardensAvailable = await hasGardens(page);
-      // SKIP: #338 owner:afo expiry:2026-08-17 — runtime data availability check
+      // SKIP: #338 owner:afo expiry:2026-09-17 — runtime data availability check
       test.skip(!gardensAvailable, "No gardens available for testing");
 
       // Queue work offline
@@ -343,7 +343,7 @@ test.describe("Work Submission Flows", () => {
       await page.waitForLoadState("domcontentloaded");
 
       const gardensAvailable = await hasGardens(page);
-      // SKIP: #338 owner:afo expiry:2026-08-17 — runtime data availability check
+      // SKIP: #338 owner:afo expiry:2026-09-17 — runtime data availability check
       test.skip(!gardensAvailable, "No gardens available for testing");
 
       // Navigate to work submission
@@ -403,7 +403,7 @@ test.describe("Work Submission Flows", () => {
 
       const gardensAvailable = await hasGardens(page);
       if (!gardensAvailable) {
-        // SKIP: #338 owner:afo expiry:2026-08-17 — runtime data availability check
+        // SKIP: #338 owner:afo expiry:2026-09-17 — runtime data availability check
         test.skip(true, "No gardens available for testing");
         return;
       }
